Namespace ControllerSpecHelper to Spec::Support::ControllerHelper
