+++ spec/spec_helper.rb
@@ -53,7 +53,7 @@ RSpec.configure do |config|
   config.include ViewSpecHelper, :type => :view
   config.include UiConstants,    :type => :view
 
-  config.include ControllerSpecHelper, :type => :controller
+  config.include Spec::Support::ControllerHelper, :type => :controller
   config.include UiConstants,          :type => :controller
   config.include Spec::Support::AuthHelper, :type => :controller
